Добавлена структура папок в pages
