--- src/pages/IngredientInfo.js
@@ -1,12 +1,0 @@
-import IngredientDetails from "../components/IngredientDetails/IngredientDetails";
-import PageLayout from "../components/PageLayout/PageLayout";
-
-const IngredientInfo = () => {
-  return (
-    <PageLayout>
-      <IngredientDetails />
-    </PageLayout>
-  );
-};
-
-export default IngredientInfo;
