--- src/pages/index.js
@@ -1,17 +1,0 @@
-import NotFound from "./NotFound";
-import Login from "./Login";
-import Register from "./Register";
-import ForgotPassword from "./ForgotPassword";
-import ResetPassword from "./ResetPassword";
-import Profile from "./Profile";
-import IngredientInfo from "./IngredientInfo";
-
-export {
-  NotFound,
-  Login,
-  Register,
-  ForgotPassword,
-  ResetPassword,
-  Profile,
-  IngredientInfo,
-};
